Before deploy

+++ ML_Bot/db/repositories/models/models.py
@@ -1,9 +1,8 @@
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, String
 from sqlalchemy.sql.schema import ForeignKey
-from sqlalchemy.dialects.postgresql import BYTEA
 from sqlalchemy.sql.sqltypes import Boolean, DateTime
 
 Base = declarative_base()
@@ -11,7 +10,7 @@ Base = declarative_base()
 class Message(Base):
     __tablename__ = "message"
 
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     text = Column(String)
     message_metadata = relationship("MessageMetadata", back_populates="message", uselist=False)
     created_at = Column(DateTime, nullable=False)
@@ -21,12 +20,12 @@ class Message(Base):
 class MessageMetadata(Base):
     __tablename__ = "message_metadata"
 
-    id = Column(Integer, primary_key=True)
-    chat_id = Column(Integer)
-    msg_id = Column(Integer, ForeignKey("message.id"), nullable=False, index=True)
-    tg_msg_id = Column(Integer, nullable=True, index=True)
+    id = Column(BigInteger, primary_key=True)
+    chat_id = Column(BigInteger)
+    msg_id = Column(BigInteger, ForeignKey("message.id"), nullable=False, index=True)
+    tg_msg_id = Column(BigInteger, nullable=True, index=True)
     message = relationship("Message", back_populates="message_metadata")
-    user_id = Column(Integer)
+    user_id = Column(BigInteger)
     links = relationship("Link", back_populates="message_metadata")
     is_deleted = Column(Boolean)
 
@@ -34,8 +33,8 @@ class MessageMetadata(Base):
 class Link(Base):
     __tablename__ = "link"
 
-    id = Column(Integer, primary_key=True)
-    message_metadata_id = Column(Integer, ForeignKey("message_metadata.id"), nullable=False)
+    id = Column(BigInteger, primary_key=True)
+    message_metadata_id = Column(BigInteger, ForeignKey("message_metadata.id"), nullable=False)
     message_metadata = relationship("MessageMetadata", back_populates="links")
     link = Column(String, index=True)
     created_at = Column(DateTime, nullable=False)
